fix: amount is double, not float

+++ ShoppingListCore/src/commonMain/kotlin/de/evylon/shoppinglist/models/Amount.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Amount(
-    val value: Float,
+    val value: Double,
     val unit: String? = null
 ) {
     override fun toString() = "$value${unit ?: ""}"
